Teach Static Python about AutoJIT
Summary:
Static Python's vtable slots assumed that function entry points were either:

* The interpreter
* A JIT-compiled function
* A wrapper that would JIT-compile the function on the next call

It did not know about AutoJIT, which uses a different wrapper function (so the
function pointers don't match in equality checks), and assumed that after it was
called it would have a JIT-compiled function ready.

This was visible when running the Lazy Imports Analyzer in with Static Python in
D61406302 which led to a crash, because the Lazy Imports Analyzer uses
AutoJIT.  It's the first "real" user of Static Python + AutoJIT out in the wild.  We
technically have it enabled for F3 as well, but I believe none of the static
functions there are hot enough to be JIT-compiled.

This diff renames and documents functions in the Jit/entry.h API, adding
`isJitEntryFunction()` to check if a function object is currently loaded with a
JIT wrapper function.  The vtable logic in Static Python is updated to only try
to fetch the entry point for a compiled static function once the entry point is
no longer the wrapper, rather than after a single call.

Differential Revision: D61572656

fbshipit-source-id: 0ceaa0689984b60debbac5366a86ae5232817ad7