Simplify CallMethod down to VectorCall for functions
Summary:
Trying to make the HIR for function calls closer between 3.10 and 3.12.

When the JIT statically knows that a CallMethod is being used for a function and
not a method (e.g. LOAD_GLOBAL + CALL in 3.12), then we can optimize to a
VectorCall and drop the `JITRT_Call` wrapper function used by CallMethod.

Reviewed By: jbower-fb

Differential Revision: D61998178

fbshipit-source-id: e7b97aedac337cb5e6b09edf2ccc2cfa282aed17