Support compiling MAKE_FUNCTION in 3.12
Summary:
MAKE_FUNCTION in 3.10 puts the new function's qualname on the top of the stack.
In 3.11 this was changed, the qualname is no longer on the stack, it's computed
from the code object's qualname.

The way this is done in HIR now is that the qualname operand becomes optional.
I swapped the operand order because I felt like it would be weird to have it as
the first argument now.

See https://docs.python.org/3.12/library/dis.html#opcode-MAKE_FUNCTION

Reviewed By: jbower-fb

Differential Revision: D62045202

fbshipit-source-id: 9c7f7972c42025131ad498b6385f15c6fe817e5b