Fix HIR support for KW_NAMES
Summary:
Turns out the `kwnames_` field wasn't being set.  Also the `emitVariadic` method
assumes that the instruction's operands are all on the stack, this isn't true
with `kwnames_`, so the stack was always popping one extra value than it was
supposed to.

The `kwnames_` logic also asserts that the last argument of the CallMethod is
still nullptr after the stack inputs have been added, as a sanity check.  This
doesn't work as `Instr::allocate()` uses `malloc()` and so the instruction isn't
zeroed out.  Switching to `calloc()` as that'll lead to more consistent
behavior.

Reviewed By: SonicField

Differential Revision: D62419069

fbshipit-source-id: cbbc462b385fa7a53304ff34da2c9402d35fb343