Fix crash on REPL startup
Summary:
The issue was D61253815 introduced a store to the caller-saved x86_64 R12 register in JIT function prologue without restoring its value on JIT function exit. The solution here is to do this in LIR code instead. This lets the automated register management handle things and is frankly more satisfactory overall.

We now get to a REPL prompt but things are still broken - there is some kind of semi-ignored Python exception during startup and an incorrect ref-count on shutdown.

Reviewed By: alexmalyshev

Differential Revision: D62334293

fbshipit-source-id: 7135c9b11d59e1ecad82c1fc35188b36a0cd0020