Move the AllPasses optimization pass to runtime tests
Summary:
This pass calls back to the top-level Compiler API to schedule all optimization
passes to run.  This leads to a dependency cycle between the HIR optimization
layer and the top-level compiler.

The only thing that uses the AllPasses class is the runtime test fixture.  Move
it there, while fixing up some uses of `const char*` to `std::string_view` that
were getting in the way.

Reviewed By: mpage

Differential Revision: D62448602

fbshipit-source-id: 079e7b226006487b31a129a5e119112c44777902