Fix rare leak in CodeAllocatorCinder
Summary:
addCode() will align up the size of the memory block to the nearest 2 MiB, and
then stores just the pointer in the allocations_ vector.  In the destructor, all
pointers in the allocations_ vector are being assumed to have size 2 MiB.

Fix this by tracking the size of each memory block.

Reviewed By: mpage

Differential Revision: D62646719

fbshipit-source-id: 1be06589d70760e6281844f0ca4ad2853d30aa74