Add extra data to code objects
Summary:
Necessary for using Cinder's JIT with Python 3.12.  Cinder needs to know how
often functions are called to determine whether it's worth JIT-compiling them.
For the 3.10.cinder runtime we added an explicit integer field to the code
object for this purpose.  This is efficient but breaks the code object ABI.  We
got away with this in 3.10.cinder because users had to explicitly onboard onto
it.  But 3.12 will become the default runtime so we cannot make ABI-breaking
changes that will preclude users from using whatever third-party extensions and
tools they'd like.

Instead, this diff makes use of the code object "extra data" API, which attaches
arbitrary pointers onto code objects.  The way it works is that a user first
needs to request a new extra data index from the runtime, then they can use that
to get and set their pointer value from any code object.

When the Cinder extension is initialized, it will set a code watcher to attach
extra data to new code objects, and then traverse all existing functions and
attach it to their code objects too.

This doesn't work yet because we don't have the custom 3.12 interpreter loop
ready in Cinder yet, and by keeping the new extra definitions in fbcode/cinderx
we can't access them from the 3.12 runtime.  So this will have to sit pretty for
now.

In the future we could do something where we upstream the call count field, and
then make use of PEP 523 to generate a modified interpreter loop that uses it.

Reviewed By: DinoV

Differential Revision: D62465238

fbshipit-source-id: f0eff930ab941adccbb6878f143a3108d12b65b0