Move definitions around in pyjit.cpp, shrink number of namespaces
Summary:
Every time I have to work with this file I find it very hard to scan it.
There's duplicate anonymous namespaces, duplicate jit namespaces, and a
smattering of static functions thrown in as well.

Simplify things by combining into a single anonymous namespace at the top,
grouping all the global variables together inside of it, and combining into a
single jit namespace.  The order of the public functions now mostly matches
what's seen in the header.  The functions for initializing and finalizing
interned strings shouldn't be prefixed with `_PyJIT_`, they're internal to the
file.

I also tweaked the JIT compilation time stuff to not use doubles.  It should use
`std::chrono::milliseconds` which is how the values are exported.

Reviewed By: jbower-fb

Differential Revision: D62647566

fbshipit-source-id: 68650700d78ca47ac73ba1b7109da26f33fe57f6