Get perf trampolines out of pyjit.cpp
Summary:
They shouldn't be mixed in with the JIT logic, they're their own thing.  Moving
them to the top-level _cinderx-lib.cpp.

The logic here is that for functions that aren't JIT-compiled, we can optionally
try to compile a trampoline for them that will be analyzable by `perf`.

The worklist can also be narrowed to only containing function objects, no code
object are ever added to it.

Reviewed By: jbower-fb

Differential Revision: D62918064

fbshipit-source-id: c0ed54af1b935e4c53876675e68b486432a9d163