Support skipping over variable sized inline cache data in more places
Summary:
This change attempts to address the fact we were assuming in some places that we could move forward/backward in bytecode by simply adding/subtracting `sizeof(_Py_CODEUNIT)` bytes from an offset. This is no longer true in 3.11+ due to inline cache data. In fact we can only efficiently move forwards.

To fix this I've refactored some things so we are always in a position to move forwards.

The biggest change is making deopt data store the location of "cause instruction"  rather than the "next instruction".

I've also refactored some of the code around two categories of deopt:
* Guards - where we need to replay the *current* (or cause) instruction.
* Other things (generally exceptions) - where technically the opcode executed correctly so we actually need to resume in the interpreter from the *next* instruction (from the one which caused the exception).

Previously this distinction was handled in `DeoptMeta::instr_offset()`. Now we only talk about "cause instruction" all the way up to frame reification and then consider this difference in `getDeoptResumeIndex()` in `deopt.cpp`. That same function also includes some other tweaks to the resume location when we are resuming from a `YIELD_FROM`.

Some other things which changed:
* I've moved the default insertion of `Snapshot` HIR instructions to the top of the bytecode translation loop from the bottom. (More on this below).
* Computing the target for jumps and of basic block boundaries in `builder.cpp`. (These were both wrong but in a consistent way so technically they would have continued to work).
* I've pre-emptively disabled the "await" call flag in 3.12 as it looks like we won't use it. This avoids the need for a way to look ahead in the bytecode stream to ensure we're seeing the correct pattern to apply the flag.

As the relative instruction captured for deopt by snapshots and frame-state has changed this means almost all the HIR output should also change. To limit the size of this diff I've made a hacky tweak to the HIR printer + parser so they still use the token `NextInstrOffset` and adjust the value so most test data won't change for 3.10 (as the offsets are always fixed for 3.10). In the next diff I've tidied this up and that diff is mostly tests data changes.

Unfortunately, there are still rather a lot of changes in this diff primarily due to slight changes in `Snapshot` insertion. I actually think the changes here are "more" correct than before but we just never noticed. Consider this example:

```
def x(a):
  a = None
  print()
```
Before this had bytecode and final HIR:
```
  3           0 LOAD_CONST               0 (None)
              2 STORE_FAST               0 (a)

  4           4 LOAD_GLOBAL              0 (print)
...

JIT: fbcode/cinderx/Jit/compiler.cpp:124 -- Optimized HIR for __main__:x:
fun __main__:x {
  bb 0 {
    v5:Object = LoadArg<0; "a">
    v6:NoneType = LoadConst<NoneType>
    v8:OptObject = LoadGlobalCached<0; "print">
    v9:MortalObjectUser[builtin_function_or_method:print:0x50800006c840] = GuardIs<0x50800006c840> v8 {
      Descr 'LOAD_GLOBAL: print'
      LiveValues<3> b:v5 b:v6 b:v8
      FrameState {
        NextInstrOffset 0
        Locals<1> v5
      }
    }
...
```
This looks wrong to me. The `NextInstrOffset` is 0 and because this is a guard we need to replay the instruction that failed, so will end up subtracting 2 bytes from `NextInstrOffset` in the former implementation of `DeoptMeta::instr_offset()` to give us an offset of -2, aka an index of -1. We would then have further subtracted 1 index to compute `f_lasti` in `reifyFrameImpl()` and that would give us an `f_lasti` of -2 which is invalid.

With these changes we now see `NextInstrOffset` as 2 i.e. a cause offset of 0, which will translate to a replay index of 0 in the new `getDeoptResumeIndex()` and `f_lasti` of -1 which *is* valid. Note even in this case we're technically resuming from the `LOAD_CONST` instruction, not the `LOAD_GLOBAL` instruction which had the potentially failing guard. However, this may technically be correct as the prior `LOAD_CONST` and `STORE_FAST` are "replayable". I don't think this logic has changed with this diff - I think it's captured in `should_snapshot()` in `builder.cpp`. I'm a bit dubious of it's correctness but I think that's a different issue from this diff.

Finally, in some cases we a similar thing happening to the `NextInstrOffset` for `RunPeriodicTasks` HIR instructions. I haven't looked at these in much detail but it looks like they have the same issue as the guards above (and are now "more" correct).

Reviewed By: alexmalyshev

Differential Revision: D63002410

fbshipit-source-id: c35f382a234abad08b1e135388aab0d0b63ae474