Stop the JIT from recursing
Summary:
If a function `A` is trying to call a Static Python function `B`, then `A` can
emit a direct x86-64 call to `B` if `B` is already compiled.  The compiler tries
to encourage this by preloading functions along with any Static Python functions
that they invoke (see `preloadFuncAndDeps`), and then when a function is being
compiled it'll try to recurse and compile those invoke targets as well.  This
makes it harder to reason about the JIT's behavior and it puts the pipeline into
an awkward dependency cycle where suddenly the HIR builder needs to reach back
to the top-level compiler API.

The same result can be achieved by capturing the functions preloaded by
`preloadFuncAndDeps` into a batch, and then compiling those one by one.  As long
as it's done in reverse post-order then the direct call optimization will still
kick in.  When the JIT goes to compile a static call from `A` to `B`, `B` should
already be compiled.

There's two gotchas here:

1) When preloading a function we also want to preload all of its called
functions in case we'd want to inline them.  However these functions should not
be _compiled_.  We only want to compile the invoked Static Python functions.

2) Currently the JIT has a recursion limit of 10.  That's a _depth_ limit and it
only applies to compilation, not preloading.  So we're compiling more than that
and preloading way way more than that.  This diff changes the preload traversal
from DFS to BFS and caps the total number of preloads for a single function to 100.
I think BFS should be "fairer", plus we do want a limit on preloads so that we don't
end up traversing the entire call graph.  This limit is configurable via a CLI flag or
an environment variable.

Reviewed By: DinoV

Differential Revision: D63478317

fbshipit-source-id: 45bdaf0b7acd25c5b5f0ed324d3a72e4fe2a6039