Bytecode refactors for D63002410
Summary:
D63002410 had a lot of changes and ended up getting reverted.  This diff tries
to collect all the various jit::Bytecode changes and other refactors in it that
shouldn't affect the behavior of the JIT.  Trying to make the next iteration of the
diff easier to review + land.

There's a few new changes I threw in though:

* BytecodeInstruction now keeps track of the PyCodeObject.  I find this makes it
  easier to work with as it'll be able to compute the next instruction offset
  correctly for 3.12 (future diff) without extra arguments.

* The BytecodeInstruction no longer stores the current opcode as it can be
  computed with the code object and the index.

Reviewed By: yuqiw4

Differential Revision: D63553454

fbshipit-source-id: 9fff644a5a4fafd4768f4b6f675c731a8be89d61