Disentangle line number table and code generation
Summary: Currently the LineAddrTable does more than just handle the generation of the line number table - it also handles the generation of the underlying bytecode! This just decouples these two factors to make LineAddrTable serve a single purpose. That'll let us have an independent algorithm which handles the newer format.

Reviewed By: alexmalyshev

Differential Revision: D62902841

fbshipit-source-id: 7c97f77f49344b2b78dccc51addf3b28610938ee