Track detailed line information internally in pyassem
Summary:
We're now receiving but throwing away the detailed line information in the assembler. This updates the assembler and downstream consumers to hold onto the detailed position information and to use that internally.

Places that were checking for -1 now can compare to `INVALID_LOCATION`, although the flow graph optimizer is an exception due to circular imports. Many existing call sites can also just use the fallback property to access the line number.

There's various annotation updates in here to help Pyre be more responsive to verifying the code changes.

Reviewed By: SonicField

Differential Revision: D62892994

fbshipit-source-id: 6c71076dcb59bcb991464a3ea8c8f4fd48856254