Initialize lnotab and insts in __init__
Summary:
The assembler has these weird semi-stateful `lnotab` and `insts` variables. They don't get explicitly initialized in `__init__` and instead get created during the lifetime of the compiler. That confounds Pyre a little bit as we actually add type annoations to the code which cause functions to be checked but it also makes the code harder to understand.

Ultimately I'm going to remove `lnotab` entirely (see D62902843) but this keeps Pyre a little bit happier as type annotations start showing up in this stack.

Reviewed By: SonicField

Differential Revision: D62902844

fbshipit-source-id: ce6f4f8f416a673b5efb3e5cb94615874abdaf44