3.12 dis_stable fixes
Summary:
I noticed in the 3.12 full test run output we had some errors when using `dis_stable`. The code object has changed (at it frequently does) so calling the constructor and passing in all of the args doesn't work. We can use `replace` which is more stable.

We also use an internal API for disassemblying the bytes which has changed, this updates us to work w/ the changes.

Reviewed By: alexmalyshev

Differential Revision: D63153150

fbshipit-source-id: d347b81e532f9a064b2b47eb5e7f39cef88f8bb3