Emit 3.12 line table format
Summary:
We now can just have the 3.12 compiler use it's own line table helper class, and generate the new format of the position data.

This is just replicating the functionality from the 3.12 compiler in Python. There's one spot where we're copying a node but not setting the line number info on it which causes the 3.12 compiler to fail, so we now copy all of the values.

We still need lots of updates to feed in the correct node when generating code though.

Reviewed By: alexmalyshev

Differential Revision: D62902842

fbshipit-source-id: e6fbe4e2337754f58cd0998ac4318b9a2d61d8a5