Make line number and byte code generation less stateful
Summary: With the `LineAddrTable` no longer holding onto the code we don't need to have this as a stateful piece of the compiler. We just have 2 passes, one to generate the byte code data as `bytes` and similar one to generate the line table info. Let's just create those and hand them off to the byte code constructor.

Reviewed By: alexmalyshev

Differential Revision: D62902843

fbshipit-source-id: 4c6db2a4a57bef1870f68076044f203596428249