Make emit not take a location
Summary: Current `PyFlowGraph.emit` can take a location, but we only use it to pass an invalid location. We're probably going to some times need to pass location for 3.12 but not for 3.10, so this makes it more explicit when we want to pass a location, and then 3.10 can have an API which drops the location.

Reviewed By: alexmalyshev

Differential Revision: D62989004

fbshipit-source-id: 3477d2911a0e2b4fe1d00f04c0c40f4cef8150e9