Don't use ThreadedCompileSerialize to guard g_compile_workers_retries
Summary:
g_compile_workers_retries should be an atomic like g_compile_workers_attempted.
Fix the comment on ThreadedCompileContext::retryUnit() which said the lock
already had to be held, this is not true.

We could go further here and optimize the atomic operations to use tighter
memory orders, but that's for a future diff.

Reviewed By: mpage

Differential Revision: D63745898

fbshipit-source-id: fc70887ebae1d341691fdcd6c509f1b0d85d5603