Make deopt-metadata store "cause" rather than "next" instruction
Summary:
This is a redo of the bits of D63002410 which haven't yet  re-landed. Particularly, it changes to the deopt metadata to be in terms of the "cause" instruction rather than the "next" instruction, eliminating the need to look backwards through full bytecode + inline cache data.

As with the previous change, this changes the logic to insert HIR `Snapshot` instructions so it's now at the top of the abstract interpreter loop rather than the bottom.

Reviewed By: alexmalyshev

Differential Revision: D64360379

fbshipit-source-id: 4dc5a9fc7e3c223f0a8c3e8dab3604684d734375