Rename borrowed.template.c to borrowed.c.template
Summary:
Our setup.py script will try to pull it in and build it which will fail, because
the script recurses through cinderx/ and finds everything that looks like an
implementation file.

I can tweak setup.py, or I tweak this file's name.  I prefer to tweak the name
because it isn't a "proper" .c file that we compile and then link into our
build, it's just a template.  So it feels odd to keep it as a .c file in our
source directory.

The downside is that we will lose highlighting for this file by default.
That's easily solvable for vim/emacs users through comment directives, but I
can't find any such equivalent for VS Code.  I imagine we could create a VS Code
project configuration for cinderx/ but that's a little beyond me.

Reviewed By: SonicField

Differential Revision: D64495095

fbshipit-source-id: 1fff6761cf14d761c9de675c51347976fbc71cbf