Fix async generator
Summary: When defining a generator we're emitting a `CALL_FUNCTION` which is no longer in 3.12. We have `emit_decorator_call` which we use for doing single argument calls, so we can just re-use that. This is relying on the fact that in 3.12 `CALL` has been generalized to always take the 1st argument or `self`.

Reviewed By: alexmalyshev

Differential Revision: D64269470

fbshipit-source-id: abfd0dc30743b53f6886d90dda338c534a577874