Relax override checks for descriptors with default values
Summary:
We currently have some tests failing due to mismatched error messages and because this check is stricter than they expect. We should still allow the override when we have the pattern:

```
class C:
    x: int = 42

class D(C):
    property
     def x(self) -> int:
        return 42
```

In this case the base class declares a slot, but it's actually a typed descriptor with value, and the field is not accessed directly.

Reviewed By: pilleye

Differential Revision: D64726776

fbshipit-source-id: bf2f1f04e4748cf2d5ef572fa4ce852b9861019c