Allow subclasses for primitive ints
Summary:
This issue was reported externally. If you do `int64(0)` and pass it to something expecting an `int64` but that isn't being invoked statically the call fails, we pass a subclass of `int` and that fails the type check.

We should accept subclasses of `int` and `float` as arguments to primitive functions, and treat them as the primitive value.

Reviewed By: alexmalyshev

Differential Revision: D64783985

fbshipit-source-id: 87b1a710eb2d1ba270f61c2443f484f84e250a75