Fix "for" code generation to use END_FOR
Summary:
This brings back D64269468 unmodified.

In 3.12 the code gen for `for` has changed in that it has a `END_FOR` opcode that `FOR_ITER` pops to. This is the thing which actually pops the value from the stack rather than having `FOR_ITER` do it when the iteration ends.

Reviewed By: alexmalyshev

Differential Revision: D64720559

fbshipit-source-id: 79984f5c2d6a738c20f08675659e56a3287404a3