Fix async generator
Summary:
This is D64269470 but with a fix for decorators defined in static Python. Our baseline tests were failing and so showed up as yellow on the original stack.  D64726775 and D64726776 should get the tests to green.

The issue with the original diff is the static compiler overrode `emit_decorator_call` and would elide the decorator visiting for data classes. When we would always generate the call to the decorator function we would detect a stack underflow.

Now `emit_decorator_call` remains, and it does the `emit_call_one_arg`. When the static compiler needs to generate the call to the decorator it uses the new `emit_call_one_arg`.

When defining a generator we're emitting a `CALL_FUNCTION` which is no longer in 3.12. We have `emit_decorator_call` which we use for doing single argument calls, so we can just re-use that. This is relying on the fact that in 3.12 `CALL` has been generalized to always take the 1st argument or `self`.

Reviewed By: alexmalyshev

Differential Revision: D64722622

fbshipit-source-id: f65a8635188d9afe50ffcc1c7deee310b6fbb7f7