Support ClassDecorators on static classes
Summary:
Take 2 of D63911500. The difference between this and the previous version is that there's now a `is_nominal_type` property for types which can be referred to at runtime. We use this when making decisions to include the type checks. The previous version only did these checks for arguments and return types and now we do it for the casts we inject within a function. And `NestedFunctionClass` is now marked as not being a nominal type.

This adds basic support for supporting a user-defined class decorator on classes in static Python. This matches a pattern that's used in IG where there is a `ClassDecorator` type which uses a `TClass` type var, and an identity function which returns the original class.

To support this we directly define the `ClassDecorator` and `TClass` types in `__static__` to keep the support here limited.

The biggest part of this is that we need some minimal support for typevar's on Python functions, including properly reporting errors when things don't line up.

One other thing is that our assertion that we can't look at annotations before the 1st phase of binding is too strong. Our model used to be very focused on trying to do 1 pass to discover types, and then we could resolve the types between modules afterwards. This doesn't work with class decorators where the decorator type needs to be known while processing the decorator. But our model has also changes to be more focused on following Python's import order, so I've eliminated the assertion. We may want to consider how to more robustly handle these circular references in the future.

There's lots of limitations around this, we need to improve our return type inference for things like `f(x: T) -> list[T]`, we need to make top-level functions have unique types for allowing for checking assignability on return, we need to fix the recursive function definition problem, etc... but this gets the core infrastructure in place that makes this existing pattern work.

Reviewed By: pilleye

Differential Revision: D64699841

fbshipit-source-id: a5bd8cc89c32ae8334f2d91acb094f4412414e02