Remove unused-variable in cinderx/Jit/compiler.cpp
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: aaronwlma

Differential Revision: D65144818

fbshipit-source-id: f358e447539f1fb7ce24097455fb7f6c2ac4d96d