Remove unused-variable in cinderx/Jit/codegen/x86_64.cpp
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: aaronwlma

Differential Revision: D65144816

fbshipit-source-id: 342a8367a212052aaf20c9b9741e1e3d31e4e3d5