Fix LoadTypeMethodCache potentially using invalid type object
Summary:
In the slow path of the cache, at the very end, it uses the cached type's name
for an attribute error string.  This is wrong because the cached type might have
been invalidated, hence why the slow path is being run.  It should instead be
checking against the type it was called with, which is the `obj` argument.

This took me forever to notice because I kept looking at `LoadTypeAttrCache`,
where `type` is a local variable. But in `LoadTypeMethodCache` it's a field on
the class.

Reviewed By: pilleye

Differential Revision: D65187749

fbshipit-source-id: 16eee1bbf4b3c698ffa807650add97bcaeaf6e51