Move cold blocks to end
Summary:
Implements some of the new flow graph based optimization, specifically moving cold blocks to the end of the function: https://www.internalfb.com/code/fbsource/[195a73c7b5c6eb371186a9e801dbd4a122b03062]/third-party/python/3.13/Python/flowgraph.c?lines=2250

CPython will now move cold blocks to the end of a function. Cold blocks are determined by first marking all methods with exception handlers, discovering all of the blocks which are reachable via normal control flow, and then marking any of the blocks reachable from exception handlers that aren't warm as cold.

Our implementation is the same algorithm, but differs a little. CPython uses bits on the block to mark cold/warm/exception handler, but it's easy enough for us to just use Python's sets to track this.

CPython's block list is also a linked list so it's manipulation is a little bit easier. Our compiler is instead mostly using this `ordered_blocks` list to track the order, so we need to shuffle things around in there. Strangely we also have a linked list, but really seem to be using it.

CPython also has a marker that a basic block doesn't fall through, and we do too, but it's also something that we don't use for calculating the stack depth, instead we have a mostly hard coded list of instructions that don't terminate.

We don't yet have the phase that converts psuedo `JUMP` instructions into the appropriate forward/backwards jump, so we just hardcode the back jumps as backwards jumps.

Reviewed By: martindemello

Differential Revision: D64790856

fbshipit-source-id: 4013e866f103487195fe11f993f4b862aea1556b