Allow including blocks in sbs code test cases
Summary:
This lets us include block markers in the side by side tests. They aren't required to be present and when the default script gets generated they get emitted as comments. If we want to test them for a specific test case we can uncomment then.

This updates the `40_yield_from_async_with.py` test to include and use these as it's a bunch of complicated code gen and it makes it clearer.

Reviewed By: martindemello

Differential Revision: D65161934

fbshipit-source-id: 5bbae63325a9871b2b6dfab7811a61076ed6c0ef