Fix initial JIT perf dump memory being leaked
Summary:
This was meant to be stored to the global `g_jitdump_mmap_addr` variable, but we
ended up creating a local with the same name instead.

This only applies to the memory allocated by `openJitdumpFile()`.
When `copyJitdumpFile()` file is called it will do the right thing and
assign to the global.

Reviewed By: jbower-fb

Differential Revision: D65352260

fbshipit-source-id: a3a98a830d078d27f97a28a1c5a7f02c750b4ac3