Fix memory leak in asyncio awaiter tracking
Summary: The bug was incorrectly marking a coroutine as awaiting itself, causing a loop which could never be GC'd. This only happened when executing the bytecode `SEND_GEN` which is a specialized variant of `SEND`. This is why this issue may have gone unnoticed in tests - the bug only kicks in after the interpreter has time to adapt bytecodes.

Reviewed By: itamaro, martindemello

Differential Revision: D65434238

fbshipit-source-id: 01e7b0df51c69eb851ba27e22e05bc85dfeb6ff8