Move block analysis to visitor
Summary:
This differs pretty substantitally in 3.12 due to the changes in comprehension inlining.

We could try and thread in some various overrides in the analyzer and have the visitor override a host of things, but that will end up being pretty ugly. It's easier to just duplicate this functionality and modify it appropriately.

This is the first step in that moving the block analysis functions into the `SymbolVisitor` so that we can get a 3.12 and 3.10 versions with different implementations.

Reviewed By: martindemello

Differential Revision: D65349476

fbshipit-source-id: 68302224d365a05f55bab4253e78fbc1aa0f997e