Implement inline comprehensions like 3.12
Summary:
Implements comprehension inlining for 3.12. 3.12 uses a `LOAD_FAST_AND_CLEAR` opcode before entering the comprehension to leave the values on the stack. It then uses a `STORE_FAST_MAYBE_NULL` to pop those values off the stack and save them when the comprehension is done (and needs to do this in a try/finally).

The symbol merging is mostly the same except 3.12 supports loading cells and has some additional accomoadations to get `__class__` right.

Reviewed By: alexmalyshev

Differential Revision: D65367827

fbshipit-source-id: e4d02b7ac3859f1948c294bd36f759751c421cab