Use Ref for LoadTypeAttrCache
Summary:
Replacing the C-like code with more C++ constructs, making more fields and
methods private too.

I wanted to do for this what was done for `LoadAttrCache`, but this cache is
only filled for one specific type of attribute.  Figuring out if that attribute
was the one that was loaded would require re-implementing the entire getter in
fill() anyway.

Reviewed By: jbower-fb

Differential Revision: D65483428

fbshipit-source-id: f3c32ad6cdc68466243964a032276c169c7a3bd5