Remove unused-variable in cinderx/Interpreter/interpreter.c
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: jbower-fb

Differential Revision: D65584274

fbshipit-source-id: 4a99f05a2b201b0897123026485524f275feea3e