Fix degenerate behavior in the simplifier
Summary:
The Simplifier replaces HIR instructions, one at a time, with cheaper
alternatives.  Sometimes it replaces an instruction with multiple instructions,
sometimes even with multiple basic blocks.  It will continuously run over a
function until it can no longer make any more changes to any of its
instructions.  It goes basic block by basic block, and it computes the next
block to process before it starts working on its current block.  This can lead
to really bad behavior when it decides to generate many new blocks.

Observe: There's a function that has 10,000 LoadAttr instructions, where the
receiver is a known Type object.  Each of those LoadAttrs will get replaced with
a combination of
LoadTypeAttrCacheEntryType/LoadTypeAttrCacheEntryValue/FillTypeAttrCache, which
adds 3 more basic blocks.  But we don't process the new blocks in the same round
of the simplifier so we restart from the beginning each time.  It goes like;

```
startSimplify
  simplifyOneRound
    process(block[0])
  simplifyOneRound
    process(block[0])
    process(block[1])
    process(block[2])
    process(block[3])
  simplifyOneRound
    process(block[0])
    process(block[1])
    process(block[2])
    process(block[3])
    process(block[4])
    process(block[5])
    process(block[6])
  ...
```

but what we really want is

```
startSimplify
  simplifyOneRound
    process(block[0])
    process(block[1])
    process(block[2])
    process(block[3])
    process(block[4])
    process(block[5])
    process(block[6])
    ...
```

Reviewed By: jbower-fb

Differential Revision: D66029338

fbshipit-source-id: c9c0fcfb3aa1e2785caf6712038801171d61f67a